Add move speed by limiting user inputs
Slowed down the rate at which user inputs are read by the MainWindow
while the user is playing, to try and fix the input lag experienced when
a user holds down a movement button. Also removed the print statements
printing out the user's position in the room after every key press.